requirements.txt
Django==3.0.5
psycopg2==2.8.5
djangorestframework==3.11.0
pytz
numpy
scipy
scikit-learn
pandas
